CE_StackInterface
CodeEval Challenge---Implement a Stack Interface.  Received a score 63.792/65 based off of correctness, memory usage and time taken

DESCRIPTION:
Write a program which implements a stack interface for integers. The interface should have ‘push’ and ‘pop’ functions. Your task is to ‘push’ a series of integers and then ‘pop’ and print every alternate integer.