CE_Time To Eat
At codeeval.com, we have a cat called Kitty. She loves eating, and it seems like she can eat anything any time. We want to keep Kitty fit, so we feed her according to a schedule. 
Planning her daily meals, we need to see when Kitty will eat starting with the latest meal in the evening to the earliest morning snack. So, your task is to sort timestamps in the schedule in a reverse chronological order.

INPUT SAMPLE:

The first argument is a path to a file. Each line includes a test case: a schedule containing unsorted timestamps in HH:MM:SS format.

For example:



1
2
02:26:31 14:44:45 09:53:27
05:33:44 21:25:41
OUTPUT SAMPLE:

Sort timestamps in each schedule from the biggest to the smallest one.



1
2
14:44:45 09:53:27 02:26:31
21:25:41 05:33:44
CONSTRAINTS:

Each schedule may have from 2 to 20 timestamps.
Timestamp 23:59:59 - biggest and 00:00:01 the smallest one.
The number of test cases is 40.